End of day
Daily Crew Charges
New Weight Goal
Generate Random Event

+++ Capstone Project GD02/Assets/Scripts/Managers/GameManager.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class GameManager : MonoBehaviour
 {
-    public float currentDay;
+    public int currentDay;
 
     public static GameManager Instance;
     void Awake()
